API: Allow using Session and Basic Auth in DEBUG mode
This makes it nicer to develop permission related changes, since it
allows the API endpoints to be accessed within the DRF Web UI using the
identity of any user logged in via Django Admin.  Also testing with curl
is easier when there is no need to generate a seperate JWT token, but
username/password can be used instead.