Use User.is_general_admin for administrator checks
Convert all code that checked the user being an administrator with the
`User.is_staff` flag to use `User.is_general_admin` instead.

This allow there to be other kind of staff too, e.g. administrator of a
single unit.

Though the Purpose endpoint still uses the is_staff check, since for
that it makes sense to make all of the purposes visible for all staff,
even if they are some other kind of staff than general administrators.

Also update the tests and extend some of them to check that bare
`User.is_staff` being True is not enough to get admin privileges.